requirements.txt
apache-airflow==2.9.3
pandas==2.2.2
pyarrow==17.0.0
pyyaml==6.0.1
pytest==8.3.2